main code for recommandExercise: q
